Persistency of data started
